tools/requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
PyYAML==6.0.1
matplotlib==3.8.2
pandas==2.1.4
seaborn==0.13.0
wordcloud==1.9.3
python-dateutil==2.8.2
pathlib==1.0.1
typing==3.7.4.3 